.continuousTest

performs a Welch Two Sample t-test or a t-test with equal variance based on the input parameters

.continuousTest(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE)

Arguments

m1

the sample mean of the first group

m2

the sample mean of the second group

s1

the sample standard deviation of the first group

s2

the sample standard deviation of the second group

n1

the sample size of the first group

n2

the sample size of the second group

m0

the null value for the difference in means to be tested for. Default is 0.

equal.variance

whether or not to assume equal variance. Default is FALSE.

Value

a list with:

continuousPValue: the p-value of the test

continuousOddsRatio: odds ration calculated as meanValueCases/meanValueControls

continuousTest: the type of test performed